feat: add concept of abstract matrix
BREAKING CHANGE:
This is a non trivial change and could potentially break existing code.
There is no known backward incompatibility though.
